chore: split the getRect of taro
