+++ package.json
@@ -185,7 +185,7 @@
     "rollup-plugin-dts": "^5.3.0",
     "sass": "^1.62.0",
     "shelljs": "^0.8.5",
-    "ts-jest": "^27.1.5",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "turndown": "^7.1.2",
     "turndown-plugin-gfm": "^1.0.2",
