+++ src/packages/menuitem/menuitem.taro.tsx
@@ -13,7 +13,7 @@ import { View } from '@tarojs/components'
 import { CSSTransition } from 'react-transition-group'
 import { Check } from '@nutui/icons-react-taro'
 import { Overlay } from '../overlay/overlay.taro'
-import { getRectByTaro } from '@/utils/use-client-rect'
+import { getRectByTaro } from '@/utils/get-rect-by-taro'
 import { BasicComponent, ComponentDefaults } from '@/utils/typings'
 
 export interface OptionItem {
