+++ src/packages/range/range.taro.tsx
@@ -9,9 +9,9 @@ import React, {
 import classNames from 'classnames'
 import { View } from '@tarojs/components'
 import { useTouch } from '@/utils/use-touch'
-import { getRectByTaro } from '@/utils/use-client-rect'
 import { BasicComponent, ComponentDefaults } from '@/utils/typings'
 import { usePropsValue } from '@/utils/use-props-value'
+import { getRectByTaro } from '@/utils/get-rect-by-taro'
 
 export type RangeValue = number | number[]
